requirements.txt
playwright
requests
python-dotenv
boto3
awslambdaric